requirements.txt
Flask
Flask-Cors
gunicorn
psycopg2-binary 
sqlalchemy
python-dotenv